Refactor Search Bar and Review Page
