+++ client/src/App.tsx
@@ -1,14 +1,6 @@
 import { Outlet, Link } from 'react-router-dom';
-import React from 'react';
-
-
-
-
-
 
 function App() {
-  
-
   return (
     
     <>
